requirements.txt
# Django and Core Framework
Django==5.2
djangorestframework==3.16.0
asgiref==3.8.1

# Database
psycopg2==2.9.10
sqlparse==0.5.3

# Authentication and Security
django-rest-passwordreset==1.5.0
django-cors-headers==4.7.0

# Environment and Configuration
python-dotenv==1.1.0

# Image Processing (for ImageField)
Pillow==10.4.0

# Date/Time Utilities
python-dateutil==2.8.2

# Translation Services
googletrans==3.1.0a0

# Type Hints (for Python < 3.9)
typing_extensions==4.13.1

# Development and Testing (optional)
# pytest==8.0.0
# pytest-django==4.8.0
# coverage==7.4.1

# Production (optional)
# gunicorn==21.2.0
# whitenoise==6.6.0
